refactor: 폴더 위치 변경 및 CI/CD 환경 변수 적용
