+++ src/main/java/com/swyp6/familytravel/auth/controller/TokenController.java
@@ -1,9 +1,9 @@
-package com.auth.controller;
+package com.swyp6.familytravel.auth.controller;
 
 
-import com.auth.jwt.JwtRequestDto;
-import com.auth.jwt.JwtResponseDto;
-import com.auth.jwt.JwtTokenUtils;
+import com.swyp6.familytravel.auth.jwt.JwtRequestDto;
+import com.swyp6.familytravel.auth.jwt.JwtResponseDto;
+import com.swyp6.familytravel.auth.jwt.JwtTokenUtils;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
