+++ src/main/java/com/swyp6/familytravel/auth/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.auth.controller;
+package com.swyp6.familytravel.auth.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
