+++ src/main/java/com/swyp6/familytravel/auth/jwt/JwtResponseDto.java
@@ -1,4 +1,4 @@
-package com.auth.jwt;
+package com.swyp6.familytravel.auth.jwt;
 
 import lombok.Getter;
 import lombok.Setter;
