+++ src/main/java/com/swyp6/familytravel/auth/jwt/JwtTokenUtils.java
@@ -1,4 +1,4 @@
-package com.auth.jwt;
+package com.swyp6.familytravel.auth.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtParser;
