+++ src/main/java/com/swyp6/familytravel/auth/service/JpaUserDetailManager.java
@@ -1,8 +1,8 @@
-package com.auth.service;
+package com.swyp6.familytravel.auth.service;
 
-import com.auth.entity.CustomUserDetails;
-import com.auth.entity.UserEntity;
-import com.auth.repository.UserRepository;
+import com.swyp6.familytravel.auth.entity.CustomUserDetails;
+import com.swyp6.familytravel.auth.entity.UserEntity;
+import com.swyp6.familytravel.auth.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.userdetails.UserDetails;
