+++ src/main/resources/application-local.yaml
@@ -1,6 +1,6 @@
 spring:
   datasource:
-    url: jdbc:mysql://localhost:3306/${DB_NAME}
+    url: jdbc:mysql://${DB_IP}:3306/${DB_NAME}
     ##?createDatabaseIfNotExist=true
     driver-class-name: com.mysql.cj.jdbc.Driver
     username: ${MYSQL_USERNAME}
@@ -14,7 +14,6 @@ spring:
         format_sql: true
         dialect: org.hibernate.dialect.MySQLDialect
     show-sql: true
-
   security:
     oauth2:
       client:
@@ -39,3 +38,9 @@ spring:
 
 jwt:
   secret: ${JWT_SECRET}
+
+images:
+  upload-root: ${UPLOAD_ROOT}
+
+feed:
+  page-size: 10
