+++ src/main/resources/application-prod.yaml
@@ -1,6 +1,7 @@
 spring:
   datasource:
     url: jdbc:mysql://${DB_IP}:3306/${DB_NAME}
+    ##?createDatabaseIfNotExist=true
     driver-class-name: com.mysql.cj.jdbc.Driver
     username: ${MYSQL_USERNAME}
     password: ${MYSQL_PASSWORD}
@@ -13,6 +14,30 @@ spring:
         format_sql: true
         dialect: org.hibernate.dialect.MySQLDialect
     show-sql: true
+  security:
+    oauth2:
+      client:
+        provider:
+          kakao:
+            authorization-uri: https://kauth.kakao.com/oauth/authorize  # 인증url
+            token-uri: https://kauth.kakao.com/oauth/token  # Acess Token
+            user-info-uri: https://kapi.kakao.com/v2/user/me  # 사용자정보 url
+            user-name-attribute: id # 응답받은 정보중 사용자이름
+        registration:
+          kakao:
+            client-id: ${KAKAO_CLIENT_ID}
+            client-secret: ${KAKAO_CLIENT_SECRET}
+            redirect-uri: ${KAKAO_REDIRECT_URI}
+            authorization-grant-type: authorization_code
+            client-authentication-method: client_secret_post
+            client-name: Kakao
+            scope:
+              - profile_nickname
+              - profile_image
+              - account_email
+
+jwt:
+  secret: ${JWT_SECRET}
 
 images:
   upload-root: ${UPLOAD_ROOT}
